[TSK-276] refactor: replace individual blueprints with a single EPI blueprint for route consistency
