+++ app/routes/EPI/fornecedores.py
@@ -1,15 +1,13 @@
-from pathlib import Path
 
-from flask import Blueprint, render_template
+from flask import render_template
 from flask_login import login_required
 
 from app.forms import CadastroFornecedores
 
-template_folder = Path(__file__).parent.resolve().joinpath("templates")
-fornecedor = Blueprint("fornecedor", __name__, template_folder=template_folder)
+from . import epi
 
 
-@fornecedor.route("/fornecedores", methods=["GET"])
+@epi.route("/fornecedores", methods=["GET"])
 @login_required
 def fornecedores():
     form = CadastroFornecedores()
