+++ app/routes/EPI/grade.py
@@ -1,17 +1,15 @@
-from pathlib import Path
 
-from flask import Blueprint, abort, render_template
+from flask import abort, render_template
 from flask_login import login_required
 
 from app.forms import CadastroGrade
 from app.misc import format_currency_brl
 from app.models import GradeEPI
 
-folder_template = Path(__file__).parent.resolve().joinpath("templates")
-grade = Blueprint("grade", __name__, template_folder=folder_template)
+from . import epi
 
 
-@grade.route("/Grade")
+@epi.route("/Grade")
 @login_required
 def Grade():
     try:
