+++ app/routes/EPI/marcas.py
@@ -1,15 +1,13 @@
-from pathlib import Path
 
-from flask import Blueprint, render_template
+from flask import render_template
 from flask_login import login_required
 
 from app.forms import CadastroMarcas
 
-template_folder = Path(__file__).parent.resolve().joinpath("templates")
-marca = Blueprint("marca", __name__, template_folder=template_folder)
+from . import epi
 
 
-@marca.route("/marcas", methods=["GET"])
+@epi.route("/marcas", methods=["GET"])
 @login_required
 def marcas():
     form = CadastroMarcas()
